Extract utils module into separate crate
