+++ Cargo.toml
@@ -37,6 +37,7 @@ tokio = { version = "1.26", features = ["fs", "rt", "tracing"] }
 tokio-stream = { version = "0.1", features = ["sync", "time"] }
 tracing = "0.1"
 tracing-futures = "0.2"
+sdk-utils = { path = "utils" }
 
 frame-support = { version = "4.0.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
 frame-system = { version = "4.0.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
@@ -221,3 +222,9 @@ sp-inherents = { version = "4.0.0-dev", git = "https://github.com/subspace/subst
 sp-io = { version = "7.0.0", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
 sp-runtime = { version = "7.0.0", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
 substrate-prometheus-endpoint = { version = "0.10.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
+
+[workspace]
+resolver = "2"
+members = [
+    "utils",
+]
