+++ src/node/domains/core_payments/chain_spec.rs
@@ -8,8 +8,7 @@ use sc_service::ChainType;
 use sc_subspace_chain_specs::ExecutionChainSpec;
 use sp_core::crypto::Ss58Codec;
 use subspace_runtime_primitives::SSC;
-
-use crate::utils::chain_spec::{chain_spec_properties, get_account_id_from_seed};
+use sdk_utils::chain_spec::{chain_spec_properties, get_account_id_from_seed};
 
 /// Chain spec
 pub type ChainSpec = ExecutionChainSpec<GenesisConfig>;
